tests(compiler): move assignments integration tests to their own file
